+++ NEWS.md
@@ -9,6 +9,7 @@ Each minor release (or even commit in the repository) is generally considered a 
 - Admin user can edit disabled fields.
 - Report config errors earlier.
 - Invoice lists issue date.
+- Support allowing partially filled placeholder team members.
 
 ### Bugs fixed